Apply MR review suggestions
