+++ executors/docker/internal/networks/manager_test.go
@@ -11,15 +11,43 @@ import (
 	"github.com/stretchr/testify/mock"
 
 	"gitlab.com/gitlab-org/gitlab-runner/common"
+	"gitlab.com/gitlab-org/gitlab-runner/executors/docker/internal/labels"
 	"gitlab.com/gitlab-org/gitlab-runner/helpers/docker"
 	"gitlab.com/gitlab-org/gitlab-runner/helpers/featureflags"
 )
+
+func newDebugLoggerMock() *mockDebugLogger {
+	loggerMock := new(mockDebugLogger)
+	loggerMock.On("Debugln", mock.Anything, mock.Anything)
+
+	return loggerMock
+}
 
 func TestNewDefaultManager(t *testing.T) {
 	logger := newDebugLoggerMock()
 
 	m := NewManager(logger, nil, nil, nil)
 	assert.IsType(t, &manager{}, m)
+}
+
+func newDefaultManager() *manager {
+	b := &common.Build{
+		ProjectRunnerID: 0,
+		Runner: &common.RunnerConfig{
+			RunnerCredentials: common.RunnerCredentials{Token: "test-token"},
+		},
+		JobResponse: common.JobResponse{
+			JobInfo: common.JobInfo{
+				ProjectID: 0,
+			},
+		},
+	}
+	m := &manager{
+		logger:  newDebugLoggerMock(),
+		build:   b,
+		labeler: labels.NewLabeler(b),
+	}
+	return m
 }
 
 func addClient(manager *manager) *docker.MockClient {
