+++ executors/docker/terminal.go
@@ -55,7 +55,7 @@ func (s *commandExecutor) Connect() (terminalsession.Conn, error) {
 	// `gitlab-terminal` package. There are plans to improve this please take a
 	// look at https://gitlab.com/gitlab-org/gitlab-ce/issues/50384#proposal and
 	// https://gitlab.com/gitlab-org/gitlab-terminal/issues/4
-	timeout := s.waitForContainerTimeout
+	timeout := s.terminalWaitForContainerTimeout
 	if timeout == 0 {
 		timeout = waitForContainerTimeout
 	}
