+++ executors/docker/terminal_test.go
@@ -75,7 +75,7 @@ func TestCommandExecutor_Connect(t *testing.T) {
 					},
 					client: c,
 				},
-				waitForContainerTimeout: 1 * time.Second,
+				terminalWaitForContainerTimeout: 1 * time.Second,
 			}
 
 			if test.hasBuildContainer {
